Fix ambiguous use of CharString operator when exporting for armeabi-v7a
When exporting for other platforms/archs the CharString "[]" operator accepts ints as values. But when exporting the project for android armeabi-v7a (arm32) the compiler errors out saying there is an "ambiguous" use of the operator. I found using int64_t as seen on the godot-cpp header fixes the error and consents exporting to armeabi-v7a (tested on my 32-bit phone).